# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Playbasis API",
        "description": "API Explorer",
        "termsOfService": "",
        "contact": {
            "email": "info@playbasis.com"
        },
        "version": "1.0"
    },
    "host": "api.pbapp.net",
    "basePath": "/",
    "schemes": [
        "https"
    ],
    "paths": {
...
        "/Service/recent_point": {
            "get": {
                "tags": [
                    "Service"
                ],
                "description": "Returns recent points",
                "parameters": [
                    {
                        "name": "point_name",
                        "in": "query",
                        "description": "Player ID as used in client's website",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Index of records to start at",
                        "required": false,
                        "type": "integer",
                        "default": 0
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of results to return",
                        "required": false,
                        "type": "integer",
                        "default": 20
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/Service/recentActivities": {
            "get": {
                "tags": [
                    "Live Feed"
                ],
                "description": "Returns recent activites",
                "parameters": [
                    {
                        "name": "player_id",
                        "in": "query",
                        "description": "Player ID as used in client's website",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Index of records to start at",
                        "required": false,
                        "type": "integer",
                        "default": 0
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of results to return",
                        "required": false,
                        "type": "integer",
                        "default": 20
                    },
                    {
                        "name": "last_read_activity_id",
                        "in": "query",
                        "description": "Last activity ID that you read",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "mode",
                        "in": "query",
                        "description": "Specify all or player",
                        "required": false,
                        "type": "string",
                        "default": "all",
                        "enum": [
                            "all",
                            "player"
                        ]
                    },
                    {
                        "name": "event_type",
                        "in": "query",
                        "description": "Event type",
                        "required": false,
                        "type": "string",
                        "enum": [
                            "reward",
                            "redeem",
                            "action",
                            "level"
                        ]
                    },
                    {
                        "name": "action_name",
                        "in": "query",
                        "description": "Filter action name",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "reward_name",
                        "in": "query",
                        "description": "Filter reward name",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/Service/detailActivityFeed/{activity_id}": {
            "get": {
                "tags": [
                    "Live Feed"
                ],
                "description": "Get detail activity",
                "parameters": [
                    {
                        "name": "activity_id",
                        "in": "path",
                        "description": "Activity ID",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/Service/likeActivityFeed/{activity_id}": {
            "post": {
                "tags": [
                    "Live Feed"
                ],
                "description": "Like activity",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Access token returned by Playbasis Authentication",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "activity_id",
                        "in": "path",
                        "description": "Activity ID",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "player_id",
                        "in": "query",
                        "description": "Player ID as used in client's website",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/Service/commentActivity/{activity_id}": {
            "post": {
                "tags": [
                    "Live Feed"
                ],
                "description": "Comment on activity",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Access token returned by Playbasis Authentication",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "activity_id",
                        "in": "path",
                        "description": "Activity ID",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "player_id",
                        "in": "query",
                        "description": "Player ID as used in client's website",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "message",
                        "in": "query",
                        "description": "Comment message",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/Service/reset_point": {
            "post": {
                "tags": [
                    "Service"
                ],
                "description": "Reset point of all players",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Access token returned by Playbasis Authentication",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "point_name",
                        "in": "query",
                        "description": "Name of the point-based reward to query",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "apiKey": {
            "type": "apiKey",
            "name": "api_key",
            "in": "query"
        },
        "token": {
            "type": "apiKey",
            "name": "token",
            "in": "query"
        }
    },
    "security": [
        {
            "apiKey": [],
            "token": []
        }
    ]
}